Integrate new jose keys into existing code
